requirements.txt
pygame==2.5.2
pyyaml

# for dev
black